(chore): 유저 리스트 조회 common 빌드 수정
