+++ constant/constant.go
@@ -13,13 +13,6 @@ var (
 func GetPath() *core.InternalApi {
 	once.Do(func() {
 		instance = &core.InternalApi{
-			Patient: core.PatientPath{
-				GetPatient:    "/patient",
-				GetPatients:   "/patient/list",
-				CreatePatient: "/patient",
-				UpdatePatient: "/patient",
-				DeletePatient: "/patient",
-			},
 			Auth: core.AuthPath{
 				CreateToken:        "/login",
 				UpdateRefreshToken: "/user",
